Bug 881: "Automatic creation of release note content broken by config change"
Added a new option to allow the caller to specify where the publish location is.

Minor improvements to formatting of output.

Turned on warnings, and fixed most of those that appeared.

Greatly simplified the checking of the script arguments.

Removed code which was noted to be unnecessary, and was.

Declared some variables more closely to the code that uses them.